LayoutTests/fast/dom/Geolocation/timeout-zero-expected.txt

Tests that when timeout is zero (and maximumAge is too), the error callback is called immediately with code TIMEOUT.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true
PASS error.code is error.TIMEOUT
PASS error.message is "Timeout expired"
PASS error.UNKNOWN_ERROR is 0
PASS error.PERMISSION_DENIED is 1
PASS error.POSITION_UNAVAILABLE is 2
PASS error.TIMEOUT is 3

TEST COMPLETE
